CPM-751: Mutualise Header component
