+++ components/identifier-generator/back/src/Infrastructure/Symfony/Resources/translations/jsmessages.en_US.yml
@@ -26,3 +26,4 @@ pim_identifier_generator:
         identifier: 'Identifier'
         first_generator: 'Create your first identifier generator!'
         check_help_center: 'Check out our help center for more information.'
+        create_info: 'It is only possible to create one Identifier Generator for the moment. You will soon be able to configure multiple generators matching all your usecases.'
