+++ components/identifier-generator/front/src/feature/components/ListSkeleton.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import {Styled} from './Styled';
+
+const ListSkeleton: React.FC = () => (
+  <>
+    <Styled.SkeletonContainer>
+      <Styled.Skeleton />
+      <Styled.Skeleton />
+      <Styled.Skeleton />
+    </Styled.SkeletonContainer>
+  </>
+);
+
+export {ListSkeleton};
