+++ components/identifier-generator/front/src/feature/components/Styled.tsx
@@ -1,4 +1,5 @@
 import styled from 'styled-components';
+import {SkeletonPlaceholder} from 'akeneo-design-system';
 
 const FormContainer = styled.div`
   display: flex;
@@ -19,8 +20,22 @@ const FullPageCenteredContent = styled.div`
   }
 `;
 
+const Skeleton = styled(SkeletonPlaceholder)`
+  width: 100%;
+  height: 50px;
+`;
+
+const SkeletonContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  width: 100%;
+`;
+
 const Styled = {
   FormContainer,
+  Skeleton,
+  SkeletonContainer,
   FullPageCenteredContent,
 };
