+++ components/identifier-generator/front/src/feature/pages/CreateOrEditGeneratorPage.tsx
@@ -1,10 +1,11 @@
-import React, {useState, useCallback} from 'react';
-import {Breadcrumb, Button, Helper, TabBar} from 'akeneo-design-system';
-import {PageContent, PageHeader, PimView, useTranslate} from '@akeneo-pim-community/shared';
+import React, {useCallback, useState} from 'react';
+import {Button, Helper, TabBar} from 'akeneo-design-system';
+import {PageContent, PageHeader, useTranslate} from '@akeneo-pim-community/shared';
 import {GeneralPropertiesTab} from '../tabs';
 import {IdentifierGenerator} from '../models';
-import {Common, Styled} from '../components';
+import {Styled} from '../components';
 import {Violation} from '../validators/Violation';
+import {Header} from '../components/Header';
 
 enum Tabs {
   GENERAL,
@@ -27,29 +28,15 @@ const CreateOrEditGeneratorPage: React.FC<CreateOrEditGeneratorProps> = ({
   const translate = useTranslate();
   const [generator, setGenerator] = useState<IdentifierGenerator>(initialGenerator);
   const changeTab = useCallback(tabName => () => setCurrentTab(tabName), []);
+  const onSave = useCallback(() => mainButtonCallback(generator), [generator, mainButtonCallback]);
 
   return (
     <>
-      <Common.Helper />
-      <PageHeader>
-        <PageHeader.Breadcrumb>
-          <Breadcrumb>
-            <Breadcrumb.Step href="#">{translate('pim_title.pim_settings_index')}</Breadcrumb.Step>
-            {/*TODO Add alert when going out this page if not saved*/}
-            <Breadcrumb.Step href="#">{translate('pim_title.akeneo_identifier_generator_index')}</Breadcrumb.Step>
-          </Breadcrumb>
-        </PageHeader.Breadcrumb>
-        <PageHeader.UserActions>
-          <PimView
-            className="AknTitleContainer-userMenuContainer AknTitleContainer-userMenu"
-            viewName="pim-identifier-generator-user-navigation"
-          />
-        </PageHeader.UserActions>
+      <Header>
         <PageHeader.Actions>
-          <Button onClick={() => mainButtonCallback(generator)}>{translate('pim_common.save')}</Button>
+          <Button onClick={onSave}>{translate('pim_common.save')}</Button>
         </PageHeader.Actions>
-        <PageHeader.Title>{translate('pim_title.akeneo_identifier_generator_index')}</PageHeader.Title>
-      </PageHeader>
+      </Header>
       <PageContent>
         {validationErrors.length > 0 && (
           <Helper level="error">
